megaplan: m5b-god-module-splits-cli-20260526-1721 execute
