+++ astrid/core/task/plan_builder.py
@@ -1,0 +1,473 @@
+"""Phase 5 lifecycle verbs: start/abort/status/runs ls/next; cmd_ack lives
+in lifecycle_ack.py to keep both modules under the size budget.
+
+cmd_runs_ls (FLAG-P5-006): natural completion does not clear active_run.json
+in V1, so the lister surfaces only 'aborted' vs 'in-progress'.
+cmd_start (SD-007): does not silently invoke compile when the pre-built JSON
+manifest is missing — prints the compile recovery and returns non-zero.
+Author-test replays are the exception: they deliberately use compiled smoke
+plans even for orchestrators that normally build dynamic start plans.
+"""
+
+from __future__ import annotations
+
+import argparse
+import hashlib
+import json
+import secrets
+import shlex
+import sys
+from dataclasses import dataclass
+from datetime import UTC, datetime
+from pathlib import Path
+from typing import Any, Optional, Sequence
+
+from astrid.core.project.current_run import (
+    clear_current_run,
+    read_current_run,
+    read_current_run_state,
+    write_current_run,
+)
+from astrid.core.project.jsonio import write_json_atomic
+from astrid.core.project.paths import (
+    project_dir,
+    resolve_projects_root,
+    validate_project_slug,
+    validate_run_id,
+)
+from astrid.core.project.project import ProjectError, require_project
+from astrid.core.alias_resolver import (
+    _register_pack_aliases,
+    create_shared_alias_resolver,
+    extract_pack_aliases,
+)
+from astrid.core.project.run import resolve_required_project_timeline
+from astrid.core.project.schema import build_run_record
+from astrid.core.pack import discover_packs
+from astrid.core.session.lease import (
+    release_writer_lease,
+    write_lease_init,
+)
+from astrid.core.session.writer import writer_context_for_project
+from astrid.core.task.claim import active_claims_by_step
+from astrid.core.task.command_render import render_task_command
+from astrid.core.task.env import task_actor_env
+from astrid.core.task.events import (
+    EventLogError,
+    _run_is_complete,
+    make_plan_initialized_event,
+    make_run_aborted_event,
+    make_run_completed_event,
+    make_run_started_event,
+    make_step_awaiting_fetch_event,
+    make_step_completed_event,
+    make_step_failed_event,
+    read_events,
+)
+from astrid.core.task.gate import TaskRunGateError, peek_current_step
+from astrid.core.task.inbox import consume_inbox_entry, pending_count, scan_inbox
+from astrid.core.task.plan import (
+    STEP_PATH_SEP,
+    RepeatForEach,
+    compute_plan_hash,
+    is_attested_kind,
+    is_code_kind,
+    is_group_step,
+    is_leaf_step,
+    iter_steps_with_path,
+    load_plan,
+    step_dir_for_path,
+)
+from astrid.core.task.plan_verbs import apply_mutations
+from astrid.core.task.preamble import PROHIBITION_PREAMBLE
+from astrid.core.timeline.crud import record_contributing_run
+from astrid.core.timeline.defaults import read_project_default
+from astrid.core.timeline.paths import find_timeline_by_slug, find_timeline_slug_for_ulid
+
+from astrid.core.task.orchestrator_resolver import _canonical_orchestrator_id, _qualified_split, _resolve_packs_root
+
+
+def _print_err(msg: str) -> None:
+    print(msg, file=sys.stderr)
+
+
+def _system_exit_code(exc: SystemExit) -> int:
+    return int(exc.code) if isinstance(exc.code, int) else 2
+
+
+_AGENT_MD_TEMPLATE = """{preamble}
+
+QUALIFIED ORCHESTRATOR: {qualified_id}
+RUN ID: {run_id}
+TIMELINE ID: {timeline_id}
+
+FIRST COMMAND (Sprint 1 / T15)
+- astrid status                    # session breadcrumb; ALWAYS run first
+- astrid attach {slug}     # bind this tab to {slug} if status reports unbound
+
+RECOVERY COMMANDS
+- See next legal action:    astrid next --project {slug}
+- Acknowledge attested:     astrid ack <step> --project {slug} --decision approve [--agent <id> | --human <name>]
+- View run state:           astrid status --project {slug}
+- End the run:              astrid abort --project {slug}
+- Take over a stuck run:    astrid sessions takeover <run-id|session-id>
+- Detach the current tab:   astrid sessions detach
+
+STOP HOOK
+- The `astrid hook stop` command is the Claude Code Stop-hook entry point.
+  When wired into .claude/settings.json (see docs/HOOKS.md) it re-injects this
+  preamble and the current step on every Stop boundary so the rules above
+  stay live for the entire run. The hook is a silent no-op outside task mode.
+
+INBOX SURFACE
+- External processes (humans, scripts, other tools) signal completion of an
+  attested step by dropping a JSON file into runs/{run_id}/inbox/.
+- File shape:
+    {{
+      "step_id": "<id of the current attested step>",
+      "decision": "approve" | "retry" | "abort",
+      "evidence": {{ "<key>": "<non-empty string>", ... }},
+      "submitted_at": "<ISO 8601 timestamp>",
+      "submitted_by": "<external system or operator name>",
+      "item_id": "<optional for_each item id>"
+    }}
+- Consume-on-next: astrid next reads inbox/, validates each file against
+  the current cursor, and appends a step_attested / item_attested /
+  cursor_rewind / run_aborted event before computing the next step.
+- Agent attestations only — human-ack steps must use `astrid ack` (the
+  inbox file would be quarantined to inbox/.rejected/ otherwise).
+- WARNING: `astrid next` is state-mutating when inbox/ has files.
+"""
+
+def _generate_run_id() -> str:
+    stamp = datetime.now(UTC).strftime("%Y%m%dT%H%M%SZ")
+    return f"run-{stamp}-{secrets.token_hex(4)}"
+
+def _plan_sha256_file(path: Path) -> str:
+    # TODO(m5b): import sha256_file from astrid.core.util.hash once the
+    # remaining lifecycle.py core↔orchestrate cleanup lands.
+    digest = hashlib.sha256()
+    with path.open("rb") as handle:
+        for chunk in iter(lambda: handle.read(1024 * 1024), b""):
+            digest.update(chunk)
+    return digest.hexdigest()
+
+def _append_consume(consumes: list[dict[str, str]], path: Path | None) -> None:
+    if path is None or not path.is_file():
+        return
+    resolved = path.resolve()
+    consumes.append({"source": str(resolved), "sha256": _plan_sha256_file(resolved)})
+
+def _default_project_video(proj_root: Path) -> Path | None:
+    video = proj_root / "source.mp4"
+    if video.is_file():
+        return video
+    return next(iter(sorted(proj_root.glob("*.mp4"))), None)
+
+def _project_file(proj_root: Path, name: str) -> Path | None:
+    path = proj_root / name
+    return path if path.is_file() else None
+
+def _hype_project_inputs(
+    proj_root: Path,
+) -> tuple[Path | None, Path | None, Path | None, list[dict[str, str]]]:
+    """Resolve conventional project-local inputs for ``astrid start video_editing.hype``."""
+    video = _default_project_video(proj_root)
+    brief = _project_file(proj_root, "brief.txt")
+    theme = _project_file(proj_root, "theme.json")
+    consumes: list[dict[str, str]] = []
+    for path in (video, brief, theme):
+        _append_consume(consumes, path)
+    return video, brief, theme, consumes
+
+def _event_talks_project_inputs(
+    proj_root: Path,
+) -> tuple[Path | None, Path | None, list[dict[str, str]]]:
+    source = _default_project_video(proj_root)
+    transcript = _project_file(proj_root, "transcript.json")
+    consumes: list[dict[str, str]] = []
+    for path in (source, transcript):
+        _append_consume(consumes, path)
+    return source, transcript, consumes
+
+def _thumbnail_maker_project_inputs(
+    proj_root: Path,
+) -> tuple[Path | None, str | None, list[dict[str, str]]]:
+    source = _default_project_video(proj_root)
+    query_path = _project_file(proj_root, "query.txt")
+    query_text = None
+    if query_path is not None:
+        query_text = query_path.read_text(encoding="utf-8").strip() or None
+    consumes: list[dict[str, str]] = []
+    for path in (source, query_path):
+        _append_consume(consumes, path)
+    return source, query_text, consumes
+
+def _build_canonical_start_plan(
+    orchestrator_id: str,
+    *,
+    proj_root: Path,
+    run_dir: Path,
+    run_id: str,
+) -> tuple[dict[str, Any], list[dict[str, str]]]:
+    if orchestrator_id == "video_editing.hype":
+        from astrid.packs.video_editing.orchestrators.hype.plan_template import build_plan_v2
+
+        video, brief, theme, consumes = _hype_project_inputs(proj_root)
+        return (
+            build_plan_v2(
+                python_exec="python3",
+                run_root=run_dir,
+                source=video,
+                brief=brief,
+                theme=theme,
+                run_id=run_id,
+            ),
+            consumes,
+        )
+    if orchestrator_id == "video_editing.event_talks":
+        from astrid.packs.video_editing.orchestrators.event_talks.plan_template import build_plan_v2
+
+        source, transcript, consumes = _event_talks_project_inputs(proj_root)
+        return (
+            build_plan_v2(
+                python_exec="python3",
+                run_root=run_dir,
+                source=source,
+                transcript=transcript,
+                run_id=run_id,
+            ),
+            consumes,
+        )
+    if orchestrator_id == "video_editing.thumbnail_maker":
+        from astrid.packs.video_editing.orchestrators.thumbnail_maker.plan_template import build_plan_v2
+
+        source, query_text, consumes = _thumbnail_maker_project_inputs(proj_root)
+        return (
+            build_plan_v2(
+                python_exec="python3",
+                run_root=run_dir,
+                source=source,
+                query=query_text,
+                run_id=run_id,
+            ),
+            consumes,
+        )
+    raise ValueError(f"unsupported canonical start orchestrator: {orchestrator_id}")
+
+def cmd_start(
+    argv: Sequence[str],
+    *,
+    packs_root: Optional[Path] = None,
+    projects_root: Optional[Path] = None,
+) -> int:
+    parser = argparse.ArgumentParser(prog="astrid start", add_help=True)
+    parser.add_argument("orchestrator_id", help="qualified id <pack>.<name>")
+    parser.add_argument("--project", required=True, help="project slug")
+    parser.add_argument("--name", default=None, help="optional run id (slug-validated)")
+    parser.add_argument("--timeline", default=None, help="timeline slug")
+    try:
+        args = parser.parse_args(list(argv))
+    except SystemExit as exc:
+        return _system_exit_code(exc)
+
+    try:
+        slug = validate_project_slug(args.project)
+    except Exception as exc:
+        _print_err(f"start: {exc}")
+        return 1
+    try:
+        require_project(slug, root=projects_root)
+    except ProjectError:
+        _print_err(
+            f"start: project {slug!r} not found; "
+            f"create one with `astrid projects create {slug}`"
+        )
+        return 1
+
+    resolved_orchestrator_id = _canonical_orchestrator_id(
+        args.orchestrator_id,
+        packs_root=_resolve_packs_root(packs_root),
+    )
+
+    try:
+        pack, name = _qualified_split(resolved_orchestrator_id)
+    except ValueError as exc:
+        _print_err(f"start: {exc}")
+        return 1
+
+    if read_current_run(slug, root=projects_root) is not None:
+        _print_err(
+            f"start: active run already exists for project {slug!r}; "
+            f"recovery: astrid abort --project {slug}"
+        )
+        return 1
+
+    uses_dynamic_start_plan = resolved_orchestrator_id in {
+        "video_editing.hype",
+        "video_editing.event_talks",
+        "video_editing.thumbnail_maker",
+    }
+    if uses_dynamic_start_plan:
+        from astrid.core.task.env import is_author_test_mode
+
+        if is_author_test_mode():
+            uses_dynamic_start_plan = False
+    if not uses_dynamic_start_plan:
+        packs = _resolve_packs_root(packs_root)
+        build_path = packs / pack / "build" / f"{name}.json"
+        if not build_path.is_file():
+            _print_err(
+                f"start: compiled plan not found at {build_path}; "
+                f"recovery: astrid author compile {resolved_orchestrator_id}"
+            )
+            return 1
+
+        try:
+            compiled_payload = json.loads(build_path.read_text(encoding="utf-8"))
+        except (OSError, json.JSONDecodeError) as exc:
+            _print_err(f"start: failed to read {build_path}: {exc}")
+            return 1
+    else:
+        compiled_payload = {}
+
+    # Resolve timeline ULID (timeline_id) and slug for display.
+    timeline_id: str | None = None
+    timeline_slug: str | None = None
+    if args.timeline is not None:
+        found = find_timeline_by_slug(slug, args.timeline, root=projects_root)
+        if found is None:
+            _print_err(
+                f"start: timeline {args.timeline!r} not found in project {slug!r}"
+            )
+            return 1
+        timeline_id = found[0]
+        timeline_slug = args.timeline
+    else:
+        default_ulid = read_project_default(slug, root=projects_root)
+        if default_ulid is not None:
+            resolved_slug = find_timeline_slug_for_ulid(slug, default_ulid, root=projects_root)
+            if resolved_slug is not None:
+                timeline_id = default_ulid
+                timeline_slug = resolved_slug
+                _print_err(
+                    f"Using default timeline: {timeline_slug}. "
+                    f"Use --timeline to override."
+                )
+    if timeline_id is None:
+        try:
+            timeline_id, timeline_slug = resolve_required_project_timeline(
+                slug,
+                root=projects_root,
+            )
+        except Exception as exc:
+            _print_err(f"start: {exc}")
+            return 1
+
+    if args.name is not None:
+        try:
+            run_id = validate_run_id(args.name)
+        except Exception as exc:
+            _print_err(f"start: --name {exc}")
+            return 1
+    else:
+        run_id = _generate_run_id()
+
+    proj_root = project_dir(slug, root=projects_root)
+    proj_root.mkdir(parents=True, exist_ok=True)
+    run_dir = proj_root / "runs" / run_id
+    run_dir.mkdir(parents=True, exist_ok=True)
+
+    consumes: list[dict[str, str]] = []
+    if uses_dynamic_start_plan:
+        try:
+            compiled_payload, consumes = _build_canonical_start_plan(
+                resolved_orchestrator_id,
+                proj_root=proj_root,
+                run_dir=run_dir,
+                run_id=run_id,
+            )
+        except Exception as exc:
+            _print_err(
+                f"start: failed to build {resolved_orchestrator_id} task plan: {exc}"
+            )
+            return 1
+
+    plan_path = proj_root / "plan.json"
+    write_json_atomic(plan_path, compiled_payload)
+    write_json_atomic(run_dir / "plan.json", compiled_payload)
+
+    try:
+        plan = load_plan(plan_path)
+    except Exception as exc:
+        _print_err(f"start: compiled plan failed validation: {exc}")
+        return 1
+
+    plan_hash = compute_plan_hash(plan_path)
+
+    run_record = build_run_record(
+        slug,
+        run_id,
+        tool_id=resolved_orchestrator_id,
+        kind="orchestrator",
+        status="prepared",
+        out=run_dir,
+        argv=["start", *list(argv)],
+        metadata={"plan_hash": plan_hash},
+        timeline_id=timeline_id,
+    )
+    if consumes:
+        run_record["consumes"] = consumes
+    write_json_atomic(
+        run_dir / "run.json",
+        run_record,
+    )
+    record_contributing_run(slug, timeline_id, run_id, root=projects_root)
+
+    # Lease-first ordering: any reader that observes current_run.json is
+    # guaranteed to find a corresponding lease.json. The session id on the
+    # lease is whatever ASTRID_SESSION_ID resolves to (CLI gate enforces
+    # the session is bound before cmd_start dispatch); fall back to
+    # 'legacy' for non-CLI callers that haven't migrated yet (tests etc).
+    from astrid.core.session.binding import (
+        SessionBindingError,
+        resolve_current_session,
+    )
+
+    session_id_for_lease = "legacy"
+    try:
+        # T9 / FLAG-S1-003: pass slug for file-bound .astrid-session fallback.
+        bound = resolve_current_session(slug=slug)
+        if bound is not None:
+            session_id_for_lease = bound.id
+    except SessionBindingError:
+        session_id_for_lease = "legacy"
+    write_lease_init(
+        run_dir,
+        session_id=session_id_for_lease,
+        plan_hash=plan_hash,
+        timeline_id=timeline_id,
+    )
+    write_current_run(slug, run_id, root=projects_root)
+
+    actor = task_actor_env()
+    started_by = f"human:{actor}" if actor else None
+    with writer_context_for_project(slug, root=projects_root) as writer:
+        writer.append(make_plan_initialized_event(run_id, plan.to_dict(), plan_hash))
+        writer.append(make_run_started_event(run_id, plan_hash, started_by=started_by))
+
+    agent_md = _AGENT_MD_TEMPLATE.format(
+        preamble=PROHIBITION_PREAMBLE,
+        qualified_id=resolved_orchestrator_id,
+        run_id=run_id,
+        slug=slug,
+        timeline_id=timeline_id,
+    )
+    (run_dir / "AGENT.md").write_text(agent_md, encoding="utf-8")
+
+    print(f"started {resolved_orchestrator_id}")
+    print(f"  project:   {slug}")
+    print(f"  timeline:  {timeline_slug}")
+    print(f"  run-id:    {run_id}")
+    print(f"  plan-hash: {plan_hash}")
+    return 0
