+++ astrid/core/timeline/branch.py
@@ -26,7 +26,7 @@ from pathlib import Path
 from typing import Any
 from uuid import uuid4
 
-from astrid import timeline as timeline_contract
+from astrid.timeline.timeline_model import validate_timeline_config_for_container
 from astrid.core.project.jsonio import write_json_atomic
 from astrid.core.util.time import utc_now_seconds as utc_now_iso
 
@@ -374,7 +374,7 @@ def _validate_branch_seed_config(
     label: str,
 ) -> dict[str, Any]:
     try:
-        return timeline_contract.validate_timeline_config_for_container(config)
+        return validate_timeline_config_for_container(config)
     except Exception as exc:
         raise ProjectionError(
             event_id=event_id,
