+++ tests/test_structure_contracts.py
@@ -42,11 +42,21 @@ def test_validate_import_layering_flags_absolute_and_relative_core_pack_imports(
     assert not any("dynamic_ok.py" in violation for violation in violations)
 
 
-def test_validate_import_layering_exempts_lifecycle_path_only(tmp_path: Path) -> None:
+def test_validate_import_layering_exempts_deferred_lifecycle_split_modules_only(tmp_path: Path) -> None:
     _write(
         tmp_path,
         "astrid/core/task/lifecycle.py",
         "from astrid.orchestrate.compile import DEFAULT_PACKS_ROOT\n",
+    )
+    _write(
+        tmp_path,
+        "astrid/core/task/orchestrator_resolver.py",
+        "from astrid.orchestrate.compile import DEFAULT_PACKS_ROOT\n",
+    )
+    _write(
+        tmp_path,
+        "astrid/core/task/plan_builder.py",
+        "from astrid.packs.video_editing.orchestrators.hype.plan_template import build_plan_v2\n",
     )
     _write(
         tmp_path,
@@ -56,8 +66,8 @@ def test_validate_import_layering_exempts_lifecycle_path_only(tmp_path: Path) ->
 
     violations = validate_import_layering(tmp_path)
 
-    assert "astrid/core/task/lifecycle.py:1 imports forbidden module 'astrid.orchestrate.compile'" not in violations
     assert violations == [
+        "astrid/core/task/lifecycle.py:1 imports forbidden module 'astrid.orchestrate.compile'",
         "astrid/core/task/not_lifecycle.py:1 imports forbidden module 'astrid.orchestrate.compile'"
     ]
